frontend/requirements.txt
pyinstaller
customtkinter
python-dotenv
